Scope the Stremio addon to a profile, and stop faking success
The addon never sent x-profile-id, so every call it made leaned on the
API's single-profile fallback. Creating a second profile broke all of
them at once — /lists and the personal-recommendation catalogs as well as
the /watch and /scrobble writes — and mark-watched still returned
"Marked as watched on Cataloggy" for a write that 400'd.

Stremio has no way to say who is watching: the installed URL is the only
thing separating one installation from another, and every resource
request hangs off the base of the manifest URL. So the profile is bound
to an installation through a path prefix — /p/<uuid>/manifest.json makes
Stremio fetch /p/<uuid>/catalog/…, /p/<uuid>/subtitles/… and so on — and
the addon threads that id into x-profile-id on every API call. Routes
stay registered bare as well, resolving to the oldest profile (the same
default getDefaultProfileId picks), so installs made before this keep
working. A malformed id in the URL is refused rather than quietly served
from the default profile. Per-profile responses are cached per profile.

Independently: the mark-watched SRT now tells the truth. A failed write,
a rejected mutation token, and an unrecognised profile each say so in the
text Stremio displays, instead of falling through to the success line.

apps/addon gains a test suite (it had none) covering both halves, which
meant splitting the Fastify app out of the listen/shutdown entrypoint.
