+++ CHANGELOG.md
@@ -17,7 +17,7 @@ Format loosely follows [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - A "Sync Status" Settings section and `GET /settings/job-status` endpoint surfacing the last failure of each scheduled background job (Trakt poll/watchlist sync, episode notifications, AI recs refresh, Steam sync, scrobble cleanup), so a self-hoster without Sentry configured can still notice a silently-broken sync.
 - A visible "update available" prompt for the web PWA instead of the service worker silently swapping itself in.
 - `PATCH /profiles/:id` for renaming a profile or setting/changing/removing its PIN, plus a Settings UI to do so — previously only creating and switching profiles was possible from the app; deleting one (already supported server-side) had no UI entry point at all.
-- Test coverage for previously-untested modules: `lib/collection.ts`, `lib/watchlist.ts`, `lib/series-progress.ts`, `routes/watch.ts`, the Plex/Jellyfin webhook handlers, and the new profile-management/job-status endpoints.
+- Test coverage for previously-untested modules: `lib/collection.ts`, `lib/watchlist.ts`, `lib/series-progress.ts`, `routes/watch.ts`, the Plex/Jellyfin webhook handlers, and the new profile-management/job-status endpoints. `apps/addon` now has a test suite of its own (previously it had none), covering profile scoping and mark-watched feedback.
 
 ### Changed
 
@@ -40,6 +40,8 @@ Format loosely follows [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Fixed
 
+- The Stremio addon now identifies which profile it is acting for, instead of silently depending on the API's single-profile fallback. Creating a second profile used to break every call the addon made — catalogs, personal recommendations, series progress, and both `/watch` and `/scrobble` writes all started failing with `x-profile-id header is required`, and "Mark Watched" still reported success. A profile is bound to an addon *installation* through its URL (`/p/<profile-id>/manifest.json`, the base Stremio derives every other request from); the Settings page shows the URL for the profile you're using. URLs without the prefix keep working and resolve to the oldest profile.
+- The addon's "Mark Watched" subtitle no longer claims success when nothing was recorded: a failed write, a rejected mutation token, and an unrecognised profile each say so in the text Stremio displays.
 - `/search` no longer throws an uncaught exception (and a bare 500) when the upstream TMDB request fails.
 - Bumped a transitive `serialize-javascript` dependency (via `workbox-build`) to patch a high-severity RCE advisory ([GHSA-5c6j-r48x-rmvq](https://github.com/advisories/GHSA-5c6j-r48x-rmvq)).
 - `getDefaultWatchlist` now retries on a concurrent-create race instead of throwing a raw Prisma error (matching `getDefaultCollection`'s existing behavior).
