+++ packages/shared/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./proxy.js";
 export * from "./sentry.js";
+export * from "./uuid.js";
